+++ src/assets/icons/overview.tsx
@@ -1,6 +1,16 @@
-export default function OverviewIcon() {
+interface Prop {
+  className?: string;
+}
+export default function OverviewIcon({ className }: Prop) {
   return (
-    <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <svg
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      className={className}
+    >
       <path
         d="M17 10H19C21 10 22 9 22 7V5C22 3 21 2 19 2H17C15 2 14 3 14 5V7C14 9 15 10 17 10Z"
         stroke="#697598"