+++ src/components/molecules/m-sidebar/index.tsx
@@ -9,7 +9,7 @@ interface SideBarProps {
 const SideBar = ({ NavItems }: SideBarProps) => {
   return (
     <div className="h-full  w-64 space-y-4  overflow-y-auto  overflow-x-hidden  bg-[#FFFFFF]  shadow-lg  transition-all  duration-500 ease-in-out  lg:fixed lg:left-0 lg:top-0 ">
-      <div className="  mb-3 w-full   px-2 py-5 text-[#003EFF]">
+      <div className="  mb-3 w-full   px-2 py-5 text-blue">
         <IndicatorIcon />
       </div>
       <div className=" space-y-2 px-4">