+++ tailwind.config.js
@@ -4,11 +4,16 @@ export default {
   theme: {
     extend: {
       fontFamily: {
-        'Neue-font':['Neue Haas Grotesk Display Pro', 'sans-serif']
+        'Neue-font':['Neue Haas Grotesk Text Pro', 'sans-serif']
       },
       colors:{
         Grey: '#697598',
-        Darkgrey:'#373B47'
+        Darkgrey:'#373B47',
+        primary: '#1F1F23',
+        blue:'#003EFF',
+        black: '#1F1F23',
+        Lightgrey:'#F6F8FA',
+        grey3 :'#666F77'
       }
     },
   },