+++ CHANGELOG.md
@@ -1,2 +1,3 @@
+- dashboard page layout
 - lint,husky and tailwind configuration
 - project setup, configuration and architecture